Configure RuboCop to permit “redundant” `%q` string literals
Although the Ruby Style Guide advises against using `%q` unless a string
contains both single and double quotes [0], we’re using it specifically
to denote strings that are intended to be treated as regular expression
literals rather than Ruby strings (with Ruby string escape sequences) so
that we can avoid unnecessary double-escaping of backslashes as well as
mistakes caused by forgetting to do it.

I’m therefore disabling the `Style/RedundantPercentQ` rule [1] to allow
us to intentionally use `%q` for this special purpose.

[0] https://rubystyle.guide/#percent-q
[1] https://docs.rubocop.org/rubocop/cops_style.html#styleredundantpercentq