Configure RuboCop to permit curly brace delimiters for `%q` strings
Although the Ruby Style Guide advises against using curly braces to
delimit a `%q` string [0], its justification is that curly braces should
be reserved for delimiting `%r` regular expression literals, because…

> parentheses often appear inside regular expressions. That’s why a less
> common character with `{` is usually the best delimiter for `%r`
> literals.

In our case we are deliberately using `%q` to quote strings that are to
be interpreted as regular expressions, so this justification applies
equally well to us.

I’m therefore configuring the `Style/PercentLiteralDelimiters` rule [1]
to allow us to intentionally use `%q{}` for this special purpose.

[0] https://rubystyle.guide/#percent-literal-braces
[1] https://docs.rubocop.org/rubocop/cops_style.html#stylepercentliteraldelimiters