Convert Lexer to use StringScanner
The standard library’s `StringScanner` class [0] is designed to do this
job, i.e. advancing a scan pointer through a string and looking for
matches.

By switching to `StringScanner` we can immediately make this code a
little more explicable (e.g. `@scanner.matched` is easier to understand
than `$&`, and `@scanner.scan` removes the need to use `$'` at all),
plus it unlocks the ability to clean up the code more in future commits.

I chose to use the empty-condition `case` syntax, in defiance of RuboCop
[1], because it more closely matches the syntax we were using before
this change, and I genuinely think it’s more readable in this context
than an `if … elsif … elsif`. But I don’t feel strongly about this so
I’m happy to be persuaded otherwise!

[0] https://rubydoc.info/stdlib/strscan/StringScanner
[1] https://docs.rubocop.org/rubocop/cops_style.html#styleemptycasecondition