Avoid escaping `^` in Lexer scan pattern
`^` only has a special meaning (negation) when it appears at the start
of a character class, so by moving it later we can avoid the need to
escape it. This is a little subtle but the tests cover it well so it
seems low-risk to arrange things to be more readable even if it’s not
immediately obvious why it works.