Avoid escaping `-` in Lexer scan pattern
Again, `-` only has a special meaning (range) when it appears between
two other characters in a character class, so by moving it to the end we
can avoid the need to escape it. A little subtle again, but it makes the
pattern easier to read and the tests will complain loudly if it ever
becomes a problem.