Configure RuboCop to enforce `%Q` style for percent string literals
We’d like to use `%Q{…}` for regular expression strings which need
interpolation, but the Ruby Style Guide [0] advises that we should only
be using `%`. For our purposes it’s more useful to use `%Q` for symmetry
with `%q`, so I’ve configured the `Style/BarePercentLiterals` cop [1] to
enforce this.

[0] https://rubystyle.guide/#percent-q-shorthand
[1] https://docs.rubocop.org/rubocop/cops_style.html#stylebarepercentliterals