Add labels to capture group states
This makes it easier to see which DFA states correspond to the NFA’s
capture group states.

We have to mutate the `match_finish` label because we need to create the
state up front (so that other transitions can use it as their target
state) but we won’t know the largest label until all the other nodes’
labels have been generated. This is slightly messy but seems worth doing
to keep the labels in a sensible order.